==> analysis/06_compositionplots.Rmd <==
---
title: "06_compositionplots"
author: "Ri Desroches"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/06_compositionplots/")

```

# Setup 

## Load Packages

```{r load-packages}
pacman::p_load(tidyverse, devtools, phyloseq, patchwork, vegan,cowplot,
               install = FALSE)


```

## set seed

```{r set-seed}

set.seed(0110)

```


## Load data

```{r load-data}
# load in normalized reads
# object called scaled_unfiltered_rooted_physeq
#repeat analysis with filtered reads IF there is enough time
# otherwise just use unfiltered reads (meaning some data has NA classification at kingdom level)
load("data/05_communityanalysis/normalizedreads.RData")

```

# Community composition plotting

## Examine data, prepare for plotting

```{r examine-data}

colnames(scaled_unfiltered_rooted_physeq@sam_data)

```

Need to add group into the data that has replicate numbers. Otherwise I won't be able to plot properly

```{r adding-reps}

# get sample data in df
sample_data_df <- data.frame(scaled_unfiltered_rooted_physeq@sam_data)

# add reps
modified_sample_data <- sample_data_df %>%
  group_by(burn_timeframe, horizon) %>%
  mutate(replicate_number = row_number())

# fix row names for merge back into phyloseq object
rownames(modified_sample_data) <- modified_sample_data$names

new_sample_data <- sample_data(modified_sample_data)

rownames(new_sample_data) <- modified_sample_data$names

# merge back into phyloseq

phyloseq_with_reps <- 
  merge_phyloseq(scaled_unfiltered_rooted_physeq, new_sample_data)

           

```

Ready for plotting

## Plot phyla grouped primarily by burn timeframe

```{r plot-phyla-by-burn-timeframe}
phylum_df <-
  phyloseq_with_reps %>%
  tax_glom(taxrank = "Phylum") %>%
  transform_sample_counts(function (x) {x/sum(x)}) %>%
  psmelt() %>%
  dplyr::filter(Abundance > 0.01) 


unique(phylum_df$Phylum)


phylum_colors <- c(
  Acidobacteriota = "#fb607F",
  Actinobacteriota = "#FFC0CB",
  Bacteroidota = "#FF7F7F",
  Chloroflexi = "#E34234",
  Crenarchaeota = "#E09540",
  Firmicutes = "#FF9966",
  Myxococcota = "#FFDEAD",
  GAL15 = "#F8DE7E",
  Gemmatimonadota = "#E3F988",
  Nitrospirota = "#F5F5DC",
  Planctomycetota = "#7BA05B" ,
  Proteobacteria = "#ACE1AF",
  "RCP2-54" ="#29AB87",
  Thermoplasmatota = "#00CCCC",
  Verrucomicrobiota = "#0D98BA",
  "WPS-2" = "#3457D5")

phylum_colors

unburned_phylum_graph_whole <- 
  phylum_df %>%
  dplyr::filter(burn_timeframe=="unburned") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~horizon) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - Unburned Soil") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors) 


oneyear_phylum_graph_whole <-
  phylum_df%>%
  dplyr::filter(burn_timeframe=="1") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~horizon) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - 1 Year Post-Burn") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors)

twoyear_phylum_graph_whole <-
  phylum_df%>%
  dplyr::filter(burn_timeframe=="2") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~horizon) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - 2 Years Post-Burn") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors)

fouryears_phylum_graph_whole <-
  phylum_df%>%
  dplyr::filter(burn_timeframe=="4") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~horizon) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - 4 Years Post-Burn") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors)


combined <- oneyear_phylum_graph_whole + twoyear_phylum_graph_whole + fouryears_phylum_graph_whole + unburned_phylum_graph_whole

combined

```

## Extract just the legend for easier plotting and tiling 

```{r extract-legend}

legend <- cowplot::get_legend(
  ggplot(phylum_df, aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
    facet_grid(. ~ horizon) + 
    geom_bar(stat = "identity", color = "black") +
    scale_fill_manual(values = phylum_colors)
)

plot_grid(legend)


  


```

 Will combine all later, after analysis runs. Can't figure out an easy way to make a nice 2x2 grid of plots with the legend floating centered and relatively small on the right

## Plot phyla primarily grouped by soil horizon

```{r plot-phyla-by-horizon}

view(phylum_df)

ahorizon_plot_whole <- phylum_df %>%
  dplyr::filter(horizon=="A horizon") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~burn_timeframe) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - A Horizon") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors) 
  
ahorizon_plot_whole


bhorizon_plot_whole <- phylum_df %>%
  dplyr::filter(horizon=="B horizon") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~burn_timeframe) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - B Horizon") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors) 
  

bhorizon_plot_whole

ehorizon_plot_whole <- phylum_df %>%
  dplyr::filter(horizon=="E horizon") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(.~burn_timeframe) + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - E Horizon") + 
  theme_bw() + 
  theme(legend.position="none", axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  scale_fill_manual(values = phylum_colors) 
  
ehorizon_plot_whole

```



## Separate bars of composition plots

Filtering JUST by soil horizon, instead of making graphs for both groups - I think it is slightly easier to follow, and produces one fewer graph.

```{r own-row-phyla}

ahorizon_plot_partial <- phylum_df %>%
  dplyr::filter(horizon=="A horizon") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(Phylum~burn_timeframe, scale="free") + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - A Horizon") + 
  theme_bw() + 
  theme(legend.position="none", 
        axis.text.x = element_text(hjust = 1, vjust = 1),
        strip.text.y.right = element_text(angle = 0)) +
  scale_fill_manual(values = phylum_colors) 


ahorizon_plot_partial

bhorizon_plot_partial <- phylum_df %>%
  dplyr::filter(horizon=="B horizon") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(Phylum~burn_timeframe, scale="free") + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - B Horizon") + 
  theme_bw() + 
    theme(legend.position="none", 
        axis.text.x = element_text(hjust = 1, vjust = 1),
        strip.text.y.right = element_text(angle = 0)) +
  scale_fill_manual(values = phylum_colors) 

bhorizon_plot_partial

ehorizon_plot_partial <- phylum_df %>%
  dplyr::filter(horizon=="E horizon") %>%
  ggplot(aes(x = replicate_number, y = Abundance, fill = Phylum)) + 
  facet_grid(Phylum~burn_timeframe, scale="free") + 
  geom_bar(stat = "identity", color = "black") + 
  labs(title = "Phylum Composition - E Horizon") + 
  theme_bw() + 
    theme(legend.position="none", 
        axis.text.x = element_text(hjust = 1, vjust = 1),
        strip.text.y.right = element_text(angle = 0)) +
  scale_fill_manual(values = phylum_colors) 


ehorizon_plot_partial

# Will not group them together, because they look really ugly!!! it is too cluttered, so i  will keep them separate

```

# Abundance boxplots

## Phylum level

Plotting both Actinobacteriota and Acidobacteriota abundance

```{r actinobacteriota-bar-plots}

# use old color vector to keep consistent

timeframe_colors <- c(
  "1"="darkorange1",
  "2"="goldenrod2",
  "4"="olivedrab",
  "unburned"="darkgreen")


phylum_df %>%
  dplyr::filter(Phylum=="Acidobacteriota") %>%
  ggplot(aes(x=burn_timeframe, y=Abundance,
             fill=burn_timeframe,color=burn_timeframe)) +
  geom_boxplot(alpha=0.6) +
  geom_jitter() +
  facet_grid(.~horizon) +
  scale_color_manual(values=timeframe_colors)+
  scale_fill_manual(values=timeframe_colors)+
  labs(title="Acidobacteriota Abundance") 


phylum_df %>%
  dplyr::filter(Phylum=="Actinobacteriota") %>%
  ggplot(aes(x=burn_timeframe, y=Abundance,
             fill=burn_timeframe,color=burn_timeframe)) +
  geom_boxplot(alpha=0.6) +
  geom_jitter() +
  facet_grid(.~horizon) +
  scale_color_manual(values=timeframe_colors)+
  scale_fill_manual(values=timeframe_colors)+
  labs(title="Actinobacteriota Abundance") 




```

## Family plotting

```{r family-plotting}

family_df <-
  phyloseq_with_reps %>%
  tax_glom(taxrank = "Family") %>%
  transform_sample_counts(function (x) {x/sum(x)}) %>%
  psmelt() %>%
  dplyr::filter(Abundance > 0.01) 

# Tried making the bar plot but there are too many values, when it gets faceted into partial barplots it is very cluttered and there is a lot of information

unique(family_df$Family)

length(unique(family_df$Family))

# Look at Acintobacteriota and Acidobacteriota

family_df %>%
  dplyr::filter(Phylum == "Actinobacteriota") %>%
  ggplot(aes(x = burn_timeframe, y = Abundance, 
             fill = burn_timeframe, color = burn_timeframe)) +
  facet_wrap(.~Family, scales = "free_y") + 
  geom_boxplot(alpha = 0.6) +
  geom_jitter() + 
  labs(title = "Actinobacteriota Family Abundance") + 
  scale_color_manual(values = timeframe_colors) + 
  scale_fill_manual(values = timeframe_colors) + 
   theme(legend.position="none", 
        axis.text.x = element_text(hjust = 1, vjust = 1),
        strip.text.y.right = element_text(angle = 0))
  
# Can not easily facet with horizon due to absence of Micrococcaceae in A horizon of all samples, for some reason



family_df %>%
  dplyr::filter(Phylum == "Acidobacteriota") %>%
  ggplot(aes(x = burn_timeframe, y = Abundance, 
             fill = burn_timeframe, color = burn_timeframe)) + 
   facet_grid(Family ~ horizon, 
              scales = "free")  + 
  geom_boxplot(alpha = 0.6) +
  geom_jitter() + 
  labs(title = "Acidobacteriota Family Abundance") + 
  scale_color_manual(values = timeframe_colors) + 
  scale_fill_manual(values = timeframe_colors) + 
   theme(legend.position="none", 
        axis.text.x = element_text(hjust = 1, vjust = 1),
        strip.text.y.right = element_text(angle = 0, size=5))

# Plotting firmicutes too  because I am curious about bacillus

  
family_df %>%
  dplyr::filter(Family == "Bacillaceae") %>%
  ggplot(aes(x = burn_timeframe, y = Abundance, 
             fill = burn_timeframe, color = burn_timeframe)) + 
  facet_wrap(.~horizon, scales = "fixed") + 
  geom_boxplot(alpha = 0.6, outlier.shape=NA) +
  geom_jitter() + 
  labs(title = "Bacillaceae Family Abundance") + 
  scale_color_manual(values = timeframe_colors) + 
  scale_fill_manual(values = timeframe_colors) + 
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1),
        legend.position = "bottom")



```

Interesting that Bacillaceae is so low in A horizon for all burn time frames, but in B and E horizons it is generally highest in one year post burn soils and gradually decreases. Maybe it is one of the primary microbes that colonizes the soil after a burn and number diminishes as the soil microbiota is restored?

# Final thoughts:

It doesn't seem, based on all analyses, that prescribed burns have an impact on the soil microbial community. The PERMANOVA test results in 05_communityanalysis were significant for soil horizon, but not burn timeframe. Even our NMDS which contains multiple axes of variation combined, there was still no separation based on even simply burnt or unburnt soils. 

We did see that the E horizon seemed to be a transition point between the A and B horizons, which is slightly counter-intuitive based on horizon names but does make sense because the E horizon is nested between the upper A horizon and the slightly lower B horizon.

When we looked in this analysis at which microbes were differentially abundant, we saw a large degree of differential abundance between burn horizons, but when examining the burn timeframes within horizons we saw little to no difference in community composition. We did not see patterns of microbial succession that we expected based on species typing. If there was any succession of specific genes or metabolic capabilities it was not visible in this analysis, but simply based on species the main driver of community composition is soil horizon and not burn timeframes.




# Session info

```{r session-info}
devtools::session_info()
```
